requirements.txt
# API server
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
stripe>=11.0.0

# Core dependencies
anthropic>=0.79.0
playwright>=1.40.0
playwright-stealth>=2.0.0
beautifulsoup4>=4.14.3
browser-use>=0.11.9
firecrawl-py>=4.15.1
pandas>=3.0.0
pydantic>=2.12.5
pydantic-settings>=2.13.0
python-dotenv>=1.2.1
pyyaml>=6.0.3
requests>=2.32.5
rapidfuzz>=3.10.0

# JWT (admin upload layer)
python-jose[cryptography]>=3.3.0

# File format readers
dbfread>=2.0.7
xlrd>=2.0.1

# SMS (Twilio)
twilio>=9.0.0

# Database dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Cache / rate-limiting (server-side only — not available locally)
redis>=5.0.0

# Agents / LangGraph (Cora runtime — Phase 2B)
langgraph>=1.1.0
langgraph-checkpoint-postgres>=3.0.0
psycopg[binary]>=3.2.0
langsmith>=0.7.0

# Sandbox — in-memory Redis shim for scenario tests
fakeredis>=2.24